app/requirements.txt
# requirements.txt
openai==1.3.0  # O la versión más reciente
ebooklib==0.18
nltk==3.8.1
spacy==3.7.2
argparse==1.4.0
python-dotenv==1.0.0